Vinay Kumar Mahto/requirements.txt
streamlit
requests
PyPDF2
pdfplumber
typing-extensions
python-dateutil
regex
pdf2image
Pillow
pytesseract
opencv-python-headless==4.10.0.84
pdf2image
pdfplumber
Pillow
PyMuPDF
PyPDF2
pytesseract
python-dateutil
regex
requests
streamlit
typing-extensions